requirements.txt
python-telegram-bot>=20.7
pymongo>=4.6
python-dotenv>=1.0.0
dnspython>=2.4
APScheduler>=3.10
loguru>=0.7
